+++ cmd/hertz_migrate/internal/cli/preload.go
@@ -1,17 +1,15 @@
 package cli
 
 import (
+	"go/token"
+	"os"
+
 	"github.com/hertz-contrib/migrate/cmd/hertz_migrate/internal/logs"
 	"github.com/hertz-contrib/migrate/cmd/hertz_migrate/internal/utils"
 	"github.com/urfave/cli/v2"
-	"go/token"
 )
 
-func preload(ctx *cli.Context) (files []string, gomods []string, err error) {
-	var (
-		gofiles   []string
-		goModDirs []string
-	)
+func preload(ctx *cli.Context) (files, gomodDirs []string, err error) {
 	fset = token.NewFileSet()
 	globalArgs.IgnoreDirs = ctx.StringSlice("ignore-dirs")
 
@@ -24,25 +22,28 @@ func preload(ctx *cli.Context) (files []string, gomods []string, err error) {
 	}
 
 	if globalArgs.TargetDir != "" {
-		gofiles, err = utils.CollectGoFiles(globalArgs.TargetDir, globalArgs.IgnoreDirs)
+		files, err = utils.CollectGoFiles(globalArgs.TargetDir, globalArgs.IgnoreDirs)
 		if err != nil {
 			return
 		}
 
-		gomods, err = utils.SearchAllDirHasGoMod(globalArgs.TargetDir)
+		gomodDirs, err = utils.SearchAllDirHasGoMod(globalArgs.TargetDir)
 		if err != nil {
 			return
 		}
 
-		for _, dir := range goModDirs {
+		for _, dir := range gomodDirs {
 			wg.Add(1)
 			dir := dir
 			go func() {
 				defer wg.Done()
-				utils.RunGoGet(dir, globalArgs.HzRepo)
+				if err := utils.RunGoGet(dir, globalArgs.HzRepo); err != nil {
+					logs.Errorf("go get hertz fail, %v", err)
+					os.Exit(1)
+				}
 			}()
 		}
 		wg.Wait()
 	}
-	return gofiles, gomods, err
+	return files, gomodDirs, err
 }
