+++ cmd/hertz_migrate/internal/cli/var.go
@@ -1,0 +1,18 @@
+package cli
+
+import (
+	"go/printer"
+	"go/token"
+	"sync"
+)
+
+var (
+	globalArgs = &Args{}
+	fset       *token.FileSet
+	wg         sync.WaitGroup
+
+	printerConfig = printer.Config{
+		Mode:     printer.UseSpaces,
+		Tabwidth: 4,
+	}
+)
