+++ cmd/hertz_migrate/internal/logic/netHttp/replaceNetHttpHandler.go
@@ -15,8 +15,9 @@
 package netHttp
 
 import (
-	mapset "github.com/deckarep/golang-set/v2"
 	. "go/ast"
+
+	mapset "github.com/deckarep/golang-set/v2"
 
 	"golang.org/x/tools/go/ast/astutil"
 )
