+++ cmd/hertz_migrate/internal/logic/netHttp/replaceReqOrRespOp.go
@@ -1,8 +1,9 @@
 package netHttp
 
 import (
-	"github.com/hertz-contrib/migrate/cmd/hertz_migrate/internal/types"
 	. "go/ast"
+
+	"github.com/hertz-contrib/migrate/cmd/hertz_migrate/internal/types"
 
 	"github.com/hertz-contrib/migrate/cmd/hertz_migrate/internal/utils"
 	"golang.org/x/tools/go/ast/astutil"
