+++ cmd/hertz_migrate/internal/logic/netHttp/replaceRequestOp.go
@@ -1,9 +1,11 @@
 package netHttp
 
 import (
+	. "go/ast"
+
 	"github.com/hertz-contrib/migrate/cmd/hertz_migrate/internal/types"
 	"github.com/hertz-contrib/migrate/cmd/hertz_migrate/internal/utils"
-	. "go/ast"
+
 	"golang.org/x/tools/go/ast/astutil"
 )
